+++ examples/tetris_point.py
@@ -75,7 +75,7 @@ class Convolution(flax.linen.Module):
 class Model(flax.linen.Module):
     @flax.linen.compact
     def __call__(self, pos, edge_src, edge_dst):
-        node_feat = e3nn.IrrepsArray.ones("0e", (pos.shape[0],))
+        node_feat = e3nn.IrrepsArray.ones("0e", (pos.shape[0],), pos.dtype)
         edge_attr = e3nn.spherical_harmonics("0e + 1o + 2e", pos[edge_dst] - pos[edge_src], True)
 
         kw = dict(