+++ ChangeLog.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `IrrepsArray.astype` to cast the underlying array
 - `e3nn.flax.MultiLayerPerceptron` and `e3nn.haiku.MultiLayerPerceptron`
+- `e3nn.IrrepsArray.from_list(..., dtype)`
 
 ### Fixed
 - set dtype for all `jnp.zeros(..., dtype)` calls in the codebase
+- set dtype for all `jnp.ones(..., dtype)` calls in the codebase
 
 ### Removed
 - `e3nn.full_tensor_product` in favor of `e3nn.tensor_product`